requirements.txt
# Google ADK and dependencies
google-adk>=1.18.0
google-cloud-aiplatform==1.123.0

# OpenTelemetry for tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-gcp-trace>=1.6.0

# Environment management
python-dotenv==1.0.0

# Async utilities (for Jupyter notebook compatibility)
nest-asyncio>=1.5.0

# Testing
pytest==9.0.1
